fixed a few tests

+++ the-codefellas-main/anber/src/main/java/io/everyonecodes/anber/profilemanagement/endpoints/UserProfileEndpoint.java
@@ -3,8 +3,11 @@ package io.everyonecodes.anber.profilemanagement.endpoints;
 import io.everyonecodes.anber.profilemanagement.service.UserProfileService;
 import io.everyonecodes.anber.usermanagement.data.User;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
+import javax.persistence.PreRemove;
 import java.util.List;
 
 @RestController
@@ -37,9 +40,18 @@ public class UserProfileEndpoint {
         return userProfileService.editData(username, profileOption, input).orElse(null);
     }
 
+    @PreRemove
     @DeleteMapping("/{username}/delete")
     @Secured("ROLE_USER")
     void deleteProfile(@PathVariable String username) {
+        String name = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if(!name.equalsIgnoreCase(username)) {
+            throw new AuthenticationException("User can only delete himself") {
+
+            };
+        }
+
         userProfileService.deleteProfile(username);
     }
